Refactor Chat components and styles; remove unused files and consolidate logic
- Deleted ChatMessageList component and its styles, replaced with a new implementation in ChatBoot.
- Removed unused styles from ChatPDM and FormSelectLine components.
- Refactored ChatPDM to use a new ChatPDMView component for better separation of concerns.
- Introduced useChatPDM and useFormSelectLine hooks for managing state and logic in ChatPDM and FormSelectLine.
- Added new EquivalenceForm and ProductEquivalenceSelector components for handling product equivalence selection.
- Implemented utility functions for extracting explanations and tables from markdown.
- Updated styles for various components to improve UI consistency.
