+++ app/components/ChatBoot/ChatBoot.module.scss
@@ -1,6 +1,3 @@
 .chatBootContainer {
-  display: flex;
-  flex-direction: column;
-  background: none;
   padding-bottom: 0.5rem;
 }
