+++ app/components/ChatBoot/EquivalenceForm/ProductEquivalenceSelector.module.scss
@@ -1,0 +1,26 @@
+.ProductEquivalenceSelector {
+  display: flex;
+  flex-direction: column;
+  border: 1px solid #ccc;
+  border-radius: 0.5rem;
+  padding: 0.7rem;
+}
+
+.buttonSubmit {
+
+  padding: 10px 24px;
+  background: #1976d2;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  font-weight: 600;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background 0.2s;
+}
+
+.buttonSubmit:disabled {
+  background: #e3eaf2;
+  color: #aaa;
+  cursor: not-allowed;
+}
