+++ app/components/ChatBoot/EquivalenceForm/types.ts
@@ -1,0 +1,15 @@
+export interface AutomotivaFields {
+  nome: string;
+  caracteristicas: string;
+  referenciaAutomotiva: string;
+  marcaFabricante: string;
+}
+
+export interface IndustrialFields {
+  nomePeca: string;
+  caracteristicasInd: string;
+  referenciaInd: string;
+  marcaInd: string;
+  norma: string;
+  aplicacao: string;
+}
