+++ app/components/ChatPDM/ChatPDMView/ChatPDMView.module.scss
@@ -3,8 +3,10 @@
   align-items: center;
   gap: 8px;
   padding: 8px;
-  border-radius: 4px;
-  background-color: #f5f5f5;
+  border-radius: 8px;
+  margin-bottom: 1rem;
+  border: 1px solid #ccc;
+
 }
 
 .inputBox input {
